==> part3.ipynb <==
import pandas as pd
df = pd.read_csv("datasets/Minimum Wage Data.csv", encoding="latin")
df.to_csv("datasets/minwage.csv", encoding="utf-8")
df = pd.read_csv("datasets/minwage.csv")
df.head()

gb = df.groupby("State")
gb.get_group("Alabama").set_index("Year").head()

act_min_wage = pd.DataFrame()

for name, group in df.groupby("State"):
    if act_min_wage.empty:
        act_min_wage = group.set_index("Year")[["Low.2018"]].rename(columns={"Low.2018":name})
    else:
        act_min_wage = act_min_wage.join(group.set_index("Year")[["Low.2018"]].rename(columns={"Low.2018":name}))

act_min_wage.head()

act_min_wage.describe()

act_min_wage.corr().head()

issue_df = df[df['Low.2018']==0]

issue_df.head()
issue_df['State'].unique()

import numpy as np
# axis 1 == columns. 0,default, is for rows
min_wage_corr=act_min_wage.replace(0, np.NaN).dropna(axis=1).corr().head()

for problem in issue_df['State'].unique():
    if problem in min_wage_corr.columns:
        print("Missing something here....")


grouped_issues = issue_df.groupby("State")

grouped_issues.get_group("Alabama").head(3)

for state, data in grouped_issues:
    if data['Low.2018'].sum() != 0.0:
        print("Some data found for", state)
