==> part8-quandl.ipynb <==
import datetime
import pandas as pd
pd.core.common.is_list_like = pd.api.types.is_list_like
import pandas_datareader.data as web
import matplotlib.pyplot as plt
from matplotlib import style
import numpy as np
import quandl
import pickle
import requests

style.use('ggplot')
#api_key = quandl.ApiConfig.api_key = 'yourAPIKey'
api_key=open('datasets/key.txt','r').read()

df=quandl.get("FMAC/HPI_AK", authtoken=api_key)
df.head()

# fiddy_states=pd.read_html('https://simple.wikipedia.org/wiki/List_of_U.S._states')
# type(fiddy_states[2])

# web = requests.get("https://simple.wikipedia.org/wiki/List_of_U.S._states")
fiddy_states = pd.read_html("https://simple.wikipedia.org/wiki/List_of_U.S._states")
states = fiddy_states[0]["Abbr"]
statesToList = states['Abbr'].tolist()

for i in statesToList:
    print(i)

df=quandl.get("FMAC/HPI_AK", authtoken=api_key)
# print(df.head())
#this is a list
def state_list():
    fiddy_states=pd.read_html('https://simple.wikipedia.org/wiki/List_of_U.S._states')
    states = fiddy_states[0]["Abbr"]
    statesToList = states['Abbr'].tolist()
    return statesToList
	#this is a dataframe
	# print(fiddy_states[0])

	# this is a column:
	# print(fiddy_states[0][0])

def grab_initial_state_data():
    states=state_list()
    main_df=pd.DataFrame()
    for abbv in states[:2]:
        query = ("FMAC/HPI_"+str(abbv))
        df=quandl.get(query,authtoken=api_key)
        #df=df.pct_change()
        df.rename(columns={'NSA Value':str(abbv) + ' NSA Value' , 'SA Value' : str(abbv) + ' SA Value'}, inplace=True)
        # df[abbv]=(df[abbv]-df[abbv][0])/df[abbv][0]*100
        df[str(abbv) + ' NSA Value'] = (df[str(abbv) + ' NSA Value']-df[str(abbv) + ' NSA Value'][0]) / df[str(abbv) + ' NSA Value'][0] * 100.0
        df[str(abbv) + ' SA Value'] = (df[str(abbv) + ' SA Value']-df[str(abbv) + ' SA Value'][0]) / df[str(abbv) + ' SA Value'][0] * 100.0

        # df.columns = [str(abbv)]
        if main_df.empty:
            main_df =df
        else:
            main_df=main_df.join(df)
    pickle_out =open('datasets/fiddy_states4.pickle','wb')
    pickle.dump(main_df,pickle_out)
    pickle_out.close()
        #print(main_df.head())
grab_initial_state_data()


def HPI_Benchmark():
    df_bmark = quandl.get("FMAC/HPI_USA", authtoken=api_key)
    df_bmark.rename(columns={'NSA Value':'US' + '_NSA Value' , 'SA Value' : 'US' + '_SA Value'},inplace=True)
    df_bmark['US' + '_NSA Value'] = (df_bmark['US' + '_NSA Value']-df_bmark['US' + '_NSA Value'][0]) /                 df_bmark['US' + '_NSA Value'][0] * 100.0
    df_bmark['US' + '_SA Value'] = (df_bmark['US' + '_SA Value']-df_bmark['US' + '_SA Value'][0]) / df_bmark['US'+ '_SA Value'][0] * 100.0
    
    return df_bmark

benchmark=HPI_Benchmark()


# pandas pickling
# HPI_data.to_pickle('pickle.pickle')
# HPI_data=pd.read_pickle('pickle.pickle')
pickle_in = open('datasets/fiddy_states4.pickle','rb')
HPI_data = pickle.load(pickle_in)
# HPI_data['TX']=HPI_data['TX']*2
# print(HPI_data['TX'])

HPI_data['AK NSA Value'].head()

fig=plt.figure()
ax1=plt.subplot2grid((1,1),(0,0))

#sampling
TX1yr=HPI_data['TX NSA Value'].resample('A',how='mean') 
print(TX1yr.head())

#handling missing data
Print(HPI_data[['TX NSA Value',TX1yr]].head())
HPI_data.dropna(inplace=True)  #how='all'
HPI_data.fillna(method='ffill',inplace=True) #fills previous values

#Rolling statistics
# HPI_data['TXMA']=pd.rolling_mean(HPI_data['TX'],12)   #df['meters'].rolling(window=2, center=False).std()﻿
# HPI_data['TXSTD']=pd.rolling_std(HPI_data['TX'],12) #deptreceated
HPI_data['TX12MA'] = pd.Series.rolling(HPI_data['TX NSA Value'], window=120).mean()

HPI_data['TX12STD'] = pd.Series.rolling(HPI_data['TX NSA Value'], window=12).std()

HPI_data['TX NSA Value'].plot(ax=ax1, label="monthly HPI TX")
TX1yr.plot(ax = ax1,label="yearly HPI TX")
# benchmark.plot(ax=ax1,color='k',linewidth=10)
plt.legend(loc=4)    #.remove()
plt.show()

# To find correlation

# HPI_State_Correlation=HPI_data.corr()
# print(HPI_State_Correlation)

# print(HPI_State_Correlation.describe())
