==> part10-Mortgage.ipynb <==
import datetime
import pandas as pd
pd.core.common.is_list_like = pd.api.types.is_list_like
import pandas_datareader.data as web
import matplotlib.pyplot as plt
from matplotlib import style
import numpy as np
import quandl
import pickle
import requests
api_key=open('datasets/key.txt','r').read()

def mortgage_30y():
    df = quandl.get("FMAC/MORTG", trim_start="1975-01-01", authtoken=api_key)
    df["Value"] = (df["Value"]-df["Value"][0]) / df["Value"][0] * 100.0
    df=df.resample('M').mean()
    print(df.head())
    return df

mortgage_30y()

HPI_data = pd.read_pickle('datasets/fiddy_states4.pickle')

def HPI_Benchmark():
    df_bmark = quandl.get("FMAC/HPI_USA", authtoken=api_key)
    df_bmark.rename(columns={'NSA Value':'US' + '_NSA Value' , 'SA Value' : 'US' + '_SA Value'},inplace=True)
    df_bmark['US' + '_NSA Value'] = (df_bmark['US' + '_NSA Value']-df_bmark['US' + '_NSA Value'][0]) /                 df_bmark['US' + '_NSA Value'][0] * 100.0
    df_bmark['US' + '_SA Value'] = (df_bmark['US' + '_SA Value']-df_bmark['US' + '_SA Value'][0]) / df_bmark['US'+ '_SA Value'][0] * 100.0
    
    return df_bmark

HPI_Bench=HPI_Benchmark()

m30 = mortgage_30y()

m30.columns=['M30']

HPI = HPI_Bench.join(m30)
print(HPI.head())

print(HPI.corr())

state_HPI_M30 = HPI_data.join(m30)
print(state_HPI_M30.corr())

print(state_HPI_M30.corr()['M30'])

print(state_HPI_M30.corr()['M30'].describe())

# def sp500_data():
#    df = quandl.get("YAHOO/INDEX_GSPC", trim_start="1975-01-01", authtoken=api_key)
#    df["Adjusted Close"] = (df["Adjusted Close"]-df["Adjusted Close"][0]) / df["Adjusted Close"][0] * 100.0
#   df=df.resample('M').mean()
#    df.rename(columns={'Adjusted Close':'sp500'}, inplace=True)
#   df = df['sp500']
#    return df
def sp500_data():
    df = quandl.get("YALE/SPCOMP", trim_start="1975-01-01", authtoken=api_key)
    df.columns = ['Adjusted Close',"1","1","1","1","1","1","1","1"]
    df["Adjusted Close"] = (df["Adjusted Close"]-df["Adjusted Close"][0]) / df["Adjusted Close"][0] * 100.0
    df=df.resample('M').mean()
    df.rename(columns={'Adjusted Close':'sp500'}, inplace=True)
    df = df['sp500']
    return df

# def gdp_data():
#    df = quandl.get("BCB/4385", trim_start="1975-01-01", authtoken=api_key)
#    df["Value"] = (df["Value"]-df["Value"][0]) / df["Value"][0] * 100.0
#    df=df.resample('M').mean()
#    df.rename(columns={'Value':'GDP'}, inplace=True)
#    df = df['GDP']
#    return df
def gdp_data():
    df = quandl.get("FRED/GDP", trim_start="1975-01-01", authtoken=api_key)
    df["Value"] = (df["Value"]-df["Value"][0]) / df["Value"][0] * 100.0
    df=df.resample('M').mean()
    df.rename(columns={'Value':'GDP'}, inplace=True)
    df = df['GDP']
    return df

#def us_unemployment():
#    df = quandl.get("ECPI/JOB_G", trim_start="1975-01-01", authtoken=api_key)
#    df["Unemployment Rate"] = (df["Unemployment Rate"]-df["Unemployment Rate"][0]) / df["Unemployment Rate"][0] * 100.0
#   df=df.resample('1D').mean()
#    df=df.resample('M').mean()
#   return df
def us_unemployment():
    df = quandl.get("FRED/UNRATE", trim_start="1975-01-01", authtoken=api_key)
    df.columns = ["Unemployment Rate"]
    df["Unemployment Rate"] = (df["Unemployment Rate"]-df["Unemployment Rate"][0]) / df["Unemployment Rate"][0] * 100.0
    df=df.resample('1D').mean()
    df=df.resample('M').mean()
    return df

sp500 = sp500_data()

gdp = gdp_data()

unemployment = us_unemployment()

HPI = HPI_Bench.join([m30,sp500,gdp,unemployment])

HPI.dropna(inplace=True)

print(HPI.corr())

 HPI.to_pickle('datasets/HPI.pickle')
